SUS-4329: Extract common test code
